# openapi.yaml
components:
  schemas:
...
    ChainResponse:
      properties:
        channel:
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        entity_id:
          title: Entity Id
          type: string
        entity_type:
          const: chain
          default: chain
          title: Entity Type
          type: string
        etag:
          title: Etag
          type: string
        meta:
          additionalProperties: true
          title: Meta
          type: object
        version_id:
          title: Version Id
          type: string
      required:
      - entity_id
      - version_id
      - channel
      - etag
      - meta
      - content
      title: ChainResponse
      type: object
...
    EntityCreateRequest:
      properties:
        channel:
          default: latest
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        evolution_meta:
          anyOf:
          - $ref: '#/components/schemas/EvolutionMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/EntityMeta'
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
      required:
      - meta
      - content
      title: EntityCreateRequest
      type: object
    EntityMeta:
      properties:
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
        namespace:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace
        tags:
          items:
            type: string
          title: Tags
          type: array
        when_to_use:
          anyOf:
          - type: string
          - type: 'null'
          title: When To Use
      required:
      - name
      title: EntityMeta
      type: object
...
    EntityUpdateRequest:
      properties:
        change_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Change Summary
        channel:
          default: latest
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        evolution_meta:
          anyOf:
          - $ref: '#/components/schemas/EvolutionMeta'
          - type: 'null'
        meta:
          anyOf:
          - $ref: '#/components/schemas/EntityMeta'
          - type: 'null'
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
      required:
      - content
      title: EntityUpdateRequest
      type: object
    EvolutionMeta:
      properties:
        behavioral_descriptors:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Behavioral Descriptors
        fitness:
          anyOf:
          - type: number
          - type: 'null'
          title: Fitness
        is_valid:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Valid
        metrics:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metrics
        mutation_kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Mutation Kind
        prompt_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt Ref
      title: EvolutionMeta
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    StepResponse:
      properties:
        channel:
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        entity_id:
          title: Entity Id
          type: string
        entity_type:
          const: step
          default: step
          title: Entity Type
          type: string
        etag:
          title: Etag
          type: string
        meta:
          additionalProperties: true
          title: Meta
          type: object
        version_id:
          title: Version Id
          type: string
      required:
      - entity_id
      - version_id
      - channel
      - etag
      - meta
      - content
      title: StepResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: 'Persistent memory for CARL artifacts: steps, chains, agents, memory
    cards'
  title: GigaEvo Memory Module
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/chains:
    get:
      description: List all chains with pagination.
      operationId: list_chains_v1_chains_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ChainResponse'
                title: Response List Chains V1 Chains Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Chains
      tags:
      - chains
    post:
      description: Create a new chain entity with its first version.
      operationId: create_chain_v1_chains_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChainResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Chain
      tags:
      - chains
  /v1/chains/{chain_id}:
    delete:
      description: Soft-delete a chain.
      operationId: delete_chain_v1_chains__chain_id__delete
      parameters:
      - in: path
        name: chain_id
        required: true
        schema:
          format: uuid
          title: Chain Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Chain
      tags:
      - chains
    get:
      description: Get a chain by ID, resolving the channel to a specific version.
      operationId: get_chain_v1_chains__chain_id__get
      parameters:
      - in: path
        name: chain_id
        required: true
        schema:
          format: uuid
          title: Chain Id
          type: string
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      - in: header
        name: if-none-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-None-Match
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChainResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Chain
      tags:
      - chains
    put:
      description: Update a chain by creating a new immutable version.
      operationId: update_chain_v1_chains__chain_id__put
      parameters:
      - in: path
        name: chain_id
        required: true
        schema:
          format: uuid
          title: Chain Id
          type: string
      - in: header
        name: if-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChainResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Chain
      tags:
      - chains
...
  /v1/steps:
    get:
      description: List all steps with pagination.
      operationId: list_steps_v1_steps_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/StepResponse'
                title: Response List Steps V1 Steps Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Steps
      tags:
      - steps
    post:
      description: Create a new step entity with its first version.
      operationId: create_step_v1_steps_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Step
      tags:
      - steps
  /v1/steps/{step_id}:
    delete:
      description: Soft-delete a step.
      operationId: delete_step_v1_steps__step_id__delete
      parameters:
      - in: path
        name: step_id
        required: true
        schema:
          format: uuid
          title: Step Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Step
      tags:
      - steps
    get:
      description: Get a step by ID, resolving the channel to a specific version.
      operationId: get_step_v1_steps__step_id__get
      parameters:
      - in: path
        name: step_id
        required: true
        schema:
          format: uuid
          title: Step Id
          type: string
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      - in: header
        name: if-none-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-None-Match
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Step
      tags:
      - steps
    put:
      description: Update a step by creating a new immutable version.
      operationId: update_step_v1_steps__step_id__put
      parameters:
      - in: path
        name: step_id
        required: true
        schema:
          format: uuid
          title: Step Id
          type: string
      - in: header
        name: if-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Step
      tags:
      - steps
...